Refactor: :recycle: refactor closePageLoader
