+++ gatsby-browser.tsx
@@ -1,23 +1,13 @@
 /**Called when the initial (but not subsequent) render of Gatsby App is done on the client. */
 import type { GatsbyBrowser } from 'gatsby';
 import { RecoilRoot } from 'recoil';
+import { closePageLoader } from './src/utils/closePageLoader';
 
 import { scrollToAnchor } from './src/utils/scrollToAnchor';
 
 export const onInitialClientRender: GatsbyBrowser['onInitialClientRender'] =
   () => {
-    const loaderIframe = document.getElementById('___loader');
-    // close loading
-    if (loaderIframe) {
-      setTimeout(() => {
-        loaderIframe.style.opacity = '0';
-      }, 0);
-
-      setTimeout(() => {
-        loaderIframe.style.display = 'none';
-      }, 300);
-    }
-
+    closePageLoader();
     scrollToAnchor(window.location.hash);
   };
